Correct optionsf.path in modem.js
Corrects optionsf.path in modem.js for http docker hosts. The path incorrect here was not breaking anything in node; however is does break browserify. This is also a bug with browserify since it's http request is not mirroring node's behavior, but it is probably good to have the path be accurate anyways.

Before the path was the full href: http://localhost:4243/<path> instead of just being /<path>