add support to http request hijacking
Implements the client side of HTTP connection hijacking described in
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.21/#3-2-hijacking

This allows for a raw TCP socket to be passed back instead of using a
custom HTTP Duplex implementation for:
 - POST /containers/(id)/attach
 - POST /exec/(id)/start

It doesn't seem to be completely documented, but it appears that any
endpoint that responds with a `{{ STREAM }}` (muxed stdout/stderr) can
be upgraded to a plain TCP connection.

To remain backwards compatible, this feature uses a new 'hijack' option
for dial instead of just assuming that all stream enable requests want
this new behaviour.

It is 100% compatible with the existing demuxStream implementation.