Customize docker-modem ssh agent's connection lifetime
When implicitly using the docker-modem agent for SSH by specifying
"protocol" as a connection option, a new agent is created for each
request.  However, it is not suitable when passed as the "agent" option
to Modem(), in which case the same agent will be reused across requests.

This is because the docker-modem ssh agent shares one SSH client for all
connections of the HTTP agent whose createConnection method it
overrides.  However, the agent appears to be capable of handling
multiple exec calls in parallel.  This means that when one stream
closes, the client will be ended, disrupting any other connections still
open.  This is fine for short-lived sequential HTTP requests made with
the agent, but if we create and attach to a docker container and then
make another request (for example, to start it), the third request will
disconnect the attached streams from the attach request.

    Mon, 09 Nov 2020 19:33:40 GMT modem.ssh1 createConnection
    2020-11-09T19:33:40.013Z modem Sending: { path: /containers/create?Image=ubuntu&AttachStdin=true&AttachStdout=true&AttachStderr=true&Tty=true&Cmd=%2Fbin%2Fbash&OpenStdin=true, ... }
    Mon, 09 Nov 2020 19:33:40 GMT modem.ssh1 ready
    Mon, 09 Nov 2020 19:33:40 GMT modem.ssh1 dialed
    2020-11-09T19:33:40.251Z modem Received: {"Id":"d13f119a1a1e60b9b54d6685e3c8e6451221b68fa9baee7f5d9c9b0ac06e52b6","Warnings":[]}
    Mon, 09 Nov 2020 19:33:40 GMT modem.ssh1 close
    Mon, 09 Nov 2020 19:33:40 GMT modem.ssh1 end
    # sleep after request to createContainer
    Mon, 09 Nov 2020 19:33:42 GMT modem.ssh2 createConnection
    2020-11-09T19:33:42.258Z modem Sending: { path: /containers/d13f119a1a1e60b9b54d6685e3c8e6451221b68fa9baee7f5d9c9b0ac06e52b6/attach?stream=true&stdout=true&stderr=true&stdin=true, ... }
    Mon, 09 Nov 2020 19:33:42 GMT modem.ssh2 ready
    Mon, 09 Nov 2020 19:33:42 GMT modem.ssh2 dialed
    # sleep after request to "attach": no "end" or "close" logged because
    # "attach" holds the connection open
    Mon, 09 Nov 2020 19:33:44 GMT modem.ssh3 createConnection
    2020-11-09T19:33:44.417Z modem Sending: { path: /containers/d13f119a1a1e60b9b54d6685e3c8e6451221b68fa9baee7f5d9c9b0ac06e52b6/start, ... }
    # both streams are closed (the one for "attach" and for "start")
    Mon, 09 Nov 2020 19:33:44 GMT modem.ssh2 end
    Mon, 09 Nov 2020 19:33:44 GMT modem.ssh3 end
    Mon, 09 Nov 2020 19:33:44 GMT modem.ssh3 ready
    Mon, 09 Nov 2020 19:33:44 GMT modem.ssh3 dialed
    2020-11-09T19:33:44.964Z modem Received:
    Mon, 09 Nov 2020 19:33:44 GMT modem.ssh3 close

Make the ssh connection local to the http agent, which creates a new,
separate ssh connection each time the http agent requests a new
connection.  This allows "attach" to stay attached.