Update ssh.js to add error handling.
Was running into an issue where I couldn't catch an error if the SSH client timed out. Adding the `.on` catch enabled me to catch it from the parent application. If it needs more changes, please let me know. For now I'm just patching it with `pnpm patch` which I would prefer not to do.